+++ crates/amaru-ledger/src/rules/transaction/phase_one/mod.rs
@@ -285,8 +285,13 @@ where
             )
         })?;
 
-        debug_span!(ledger::rules::phase_one::VOTES)
-            .in_scope(|| voting_procedures::execute(context, mem::take(&mut transaction_body.votes)))?;
+        debug_span!(ledger::rules::phase_one::VOTES).in_scope(|| {
+            voting_procedures::execute(
+                context,
+                protocol_parameters.protocol_version,
+                mem::take(&mut transaction_body.votes),
+            )
+        })?;
     } else {
         debug_span!(ledger::rules::phase_one::CERTIFICATES).in_scope(|| {
             certificates::count_lovelace(context, protocol_parameters, mem::take(&mut transaction_body.certificates))